# Exercicio-main/student-subject-api/src/main/java/com/example/studentsubjectapi/model/Student.java
package com.example.studentsubjectapi.model;

public class Student {
    private String name;
    private String registrationNumber;

    public Student() {
    }

    public Student(String name, String registrationNumber) {
        this.name = name;
        this.registrationNumber = registrationNumber;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getRegistrationNumber() {
        return registrationNumber;
    }

    public void setRegistrationNumber(String registrationNumber) {
        this.registrationNumber = registrationNumber;
    }
}